move zq/ast to compiler/zq/ast
This commit moves the ast package under compiler.  The goal here is
to put all compiler related code under zq/compiler, e.g., forthcoming
PRs will create compiler/flow and compiler/semantic.
