+++ api/client/multipartwriter.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/brimsec/zq/ast"
+	"github.com/brimsec/zq/compiler/ast"
 	"github.com/brimsec/zq/pkg/iosrc"
 	"github.com/brimsec/zq/zio/ndjsonio"
 )
