+++ compiler/expr.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/brimsec/zq/ast"
+	"github.com/brimsec/zq/compiler/ast"
 	"github.com/brimsec/zq/expr"
 	"github.com/brimsec/zq/expr/agg"
 	"github.com/brimsec/zq/expr/function"
