+++ expr/boolean.go
@@ -8,7 +8,9 @@ import (
 	"regexp"
 	"regexp/syntax"
 
-	"github.com/brimsec/zq/ast"
+	//XXX this shouldn't be reaching into the AST but we'll leave it for
+	// now until we factor-in the flow-based package
+	"github.com/brimsec/zq/compiler/ast"
 	"github.com/brimsec/zq/pkg/byteconv"
 	"github.com/brimsec/zq/zng"
 )
