+++ expr/literal.go
@@ -1,7 +1,7 @@
 package expr
 
 import (
-	"github.com/brimsec/zq/ast"
+	"github.com/brimsec/zq/compiler/ast"
 	"github.com/brimsec/zq/zng"
 )
