+++ ppl/lake/index/definition.go
@@ -6,7 +6,7 @@ import (
 	"path"
 	"regexp"
 
-	"github.com/brimsec/zq/ast"
+	"github.com/brimsec/zq/compiler/ast"
 	"github.com/brimsec/zq/pkg/iosrc"
 	"github.com/segmentio/ksuid"
 )
