+++ proc/groupby/groupby_test.go
@@ -11,8 +11,8 @@ import (
 	"testing"
 
 	"github.com/brimsec/zq/api"
-	"github.com/brimsec/zq/ast"
 	"github.com/brimsec/zq/compiler"
+	"github.com/brimsec/zq/compiler/ast"
 	"github.com/brimsec/zq/driver"
 	"github.com/brimsec/zq/pkg/test"
 	"github.com/brimsec/zq/proc/groupby"
