+++ zson/zson_test.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"testing"
 
-	"github.com/brimsec/zq/ast"
+	"github.com/brimsec/zq/compiler/ast"
 	"github.com/brimsec/zq/pkg/fs"
 	"github.com/brimsec/zq/zng"
 	"github.com/brimsec/zq/zng/resolver"
